+++ ui/src/components/CuiDialog/templates/NotificationDetail/types.ts
@@ -1,0 +1,4 @@
+export interface NotificationDetailProps {
+  src: string;
+  text: string;
+}